BREAKING: drop Federation v1 support
Federation v2 was released almost 3 years ago. It is an evolution of the Federation spec to make it more powerful, flexible and easier to adopt.

It is no longer possible to create Federated Supergraphs targeting v1 composition. Users still using fed v1 shoul migrate ASAP to the fed v2.
