+++ generator/graphql-kotlin-federation/src/main/kotlin/com/expediagroup/graphql/generator/federation/FederatedSchemaGeneratorHooks.kt
@@ -16,7 +16,6 @@
 
 package com.expediagroup.graphql.generator.federation
 
-import com.apollographql.federation.graphqljava.printer.ServiceSDLPrinter.generateServiceSDL
 import com.apollographql.federation.graphqljava.printer.ServiceSDLPrinter.generateServiceSDLV2
 import com.expediagroup.graphql.generator.TopLevelObject
 import com.expediagroup.graphql.generator.annotations.GraphQLName
@@ -25,10 +24,8 @@ import com.expediagroup.graphql.generator.federation.directives.COMPOSE_DIRECTIV
 import com.expediagroup.graphql.generator.federation.directives.CONTACT_DIRECTIVE_NAME
 import com.expediagroup.graphql.generator.federation.directives.CONTACT_DIRECTIVE_TYPE
 import com.expediagroup.graphql.generator.federation.directives.EXTENDS_DIRECTIVE_NAME
-import com.expediagroup.graphql.generator.federation.directives.EXTENDS_DIRECTIVE_TYPE
 import com.expediagroup.graphql.generator.federation.directives.EXTERNAL_DIRECTIVE_NAME
 import com.expediagroup.graphql.generator.federation.directives.EXTERNAL_DIRECTIVE_TYPE
-import com.expediagroup.graphql.generator.federation.directives.EXTERNAL_DIRECTIVE_TYPE_V2
 import com.expediagroup.graphql.generator.federation.directives.FEDERATION_SPEC
 import com.expediagroup.graphql.generator.federation.directives.FEDERATION_SPEC_LATEST_URL
 import com.expediagroup.graphql.generator.federation.directives.FEDERATION_SPEC_URL_PREFIX
@@ -36,7 +33,6 @@ import com.expediagroup.graphql.generator.federation.directives.FieldSet
 import com.expediagroup.graphql.generator.federation.directives.INACCESSIBLE_DIRECTIVE_NAME
 import com.expediagroup.graphql.generator.federation.directives.INTERFACE_OBJECT_DIRECTIVE_NAME
 import com.expediagroup.graphql.generator.federation.directives.KEY_DIRECTIVE_NAME
-import com.expediagroup.graphql.generator.federation.directives.KEY_DIRECTIVE_TYPE
 import com.expediagroup.graphql.generator.federation.directives.LINK_DIRECTIVE_NAME
 import com.expediagroup.graphql.generator.federation.directives.LINK_SPEC
 import com.expediagroup.graphql.generator.federation.directives.LinkDirective
@@ -45,9 +41,7 @@ import com.expediagroup.graphql.generator.federation.directives.LinkedSpec
 import com.expediagroup.graphql.generator.federation.directives.OVERRIDE_DIRECTIVE_NAME
 import com.expediagroup.graphql.generator.federation.directives.POLICY_DIRECTIVE_NAME
 import com.expediagroup.graphql.generator.federation.directives.PROVIDES_DIRECTIVE_NAME
-import com.expediagroup.graphql.generator.federation.directives.PROVIDES_DIRECTIVE_TYPE
 import com.expediagroup.graphql.generator.federation.directives.REQUIRES_DIRECTIVE_NAME
-import com.expediagroup.graphql.generator.federation.directives.REQUIRES_DIRECTIVE_TYPE
 import com.expediagroup.graphql.generator.federation.directives.REQUIRES_SCOPE_DIRECTIVE_NAME
 import com.expediagroup.graphql.generator.federation.directives.SHAREABLE_DIRECTIVE_NAME
 import com.expediagroup.graphql.generator.federation.directives.TAG_DIRECTIVE_NAME
@@ -61,7 +55,6 @@ import com.expediagroup.graphql.generator.federation.directives.toAppliedLinkDir
 import com.expediagroup.graphql.generator.federation.directives.toAppliedPolicyDirective
 import com.expediagroup.graphql.generator.federation.directives.toAppliedRequiresScopesDirective
 import com.expediagroup.graphql.generator.federation.exception.DuplicateSpecificationLinkImport
-import com.expediagroup.graphql.generator.federation.exception.IncorrectFederatedDirectiveUsage
 import com.expediagroup.graphql.generator.federation.exception.InvalidFederatedSchema
 import com.expediagroup.graphql.generator.federation.exception.UnknownSpecificationException
 import com.expediagroup.graphql.generator.federation.execution.EntitiesDataFetcher
@@ -70,7 +63,6 @@ import com.expediagroup.graphql.generator.federation.types.ANY_SCALAR_TYPE
 import com.expediagroup.graphql.generator.federation.types.ENTITY_UNION_NAME
 import com.expediagroup.graphql.generator.federation.types.FIELD_SET_SCALAR_NAME
 import com.expediagroup.graphql.generator.federation.types.FIELD_SET_SCALAR_TYPE
-import com.expediagroup.graphql.generator.federation.types.FieldSetTransformer
 import com.expediagroup.graphql.generator.federation.types.LINK_IMPORT_SCALAR_TYPE
 import com.expediagroup.graphql.generator.federation.types.POLICY_SCALAR_TYPE
 import com.expediagroup.graphql.generator.federation.types.SCOPE_SCALAR_TYPE
@@ -91,7 +83,6 @@ import graphql.schema.GraphQLObjectType
 import graphql.schema.GraphQLScalarType
 import graphql.schema.GraphQLSchema
 import graphql.schema.GraphQLType
-import graphql.schema.SchemaTransformer
 import java.nio.file.Paths
 import kotlin.io.path.name
 import kotlin.reflect.KType
@@ -101,44 +92,22 @@ import kotlin.reflect.full.findAnnotation
  * Hooks for generating federated GraphQL schema.
  */
 open class FederatedSchemaGeneratorHooks(
-    private val resolvers: List<FederatedTypeResolver>,
-    private val optInFederationV2: Boolean = true
+    private val resolvers: List<FederatedTypeResolver>
 ) : FlowSubscriptionSchemaGeneratorHooks() {
     private val validator: FederatedSchemaValidator = FederatedSchemaValidator()
     data class LinkSpec(val namespace: String, val imports: Map<String, String>)
     private val linkSpecs: MutableMap<String, LinkSpec> = HashMap()
 
-    private val federationV2OnlyDirectiveNames: Set<String> = setOf(
-        COMPOSE_DIRECTIVE_NAME,
-        INACCESSIBLE_DIRECTIVE_NAME,
-        INTERFACE_OBJECT_DIRECTIVE_NAME,
-        LINK_DIRECTIVE_NAME,
-        OVERRIDE_DIRECTIVE_NAME,
-        SHAREABLE_DIRECTIVE_NAME
-    )
-
-    private val federatedDirectiveV1List: List<GraphQLDirective> = listOf(
-        EXTENDS_DIRECTIVE_TYPE,
-        EXTERNAL_DIRECTIVE_TYPE,
-        KEY_DIRECTIVE_TYPE,
-        PROVIDES_DIRECTIVE_TYPE,
-        REQUIRES_DIRECTIVE_TYPE
-    )
-
     // workaround to https://github.com/ExpediaGroup/graphql-kotlin/issues/1815
     // since those scalars can be renamed, we need to ensure we only generate those scalars just once
     private val fieldSetScalar: GraphQLScalarType by lazy {
-        if (optInFederationV2) {
-            FIELD_SET_SCALAR_TYPE.run {
-                val fieldSetScalarName = namespacedTypeName(FEDERATION_SPEC, this.name)
-                if (fieldSetScalarName != this.name) {
-                    return@run this.transform { it.name(fieldSetScalarName) }
-                } else {
-                    this
-                }
+        FIELD_SET_SCALAR_TYPE.run {
+            val fieldSetScalarName = namespacedTypeName(FEDERATION_SPEC, this.name)
+            if (fieldSetScalarName != this.name) {
+                return@run this.transform { it.name(fieldSetScalarName) }
+            } else {
+                this
             }
-        } else {
-            FIELD_SET_SCALAR_TYPE
         }
     }
     private val linkImportScalar: GraphQLScalarType by lazy {
@@ -186,52 +155,50 @@ open class FederatedSchemaGeneratorHooks(
         additionalInputTypes: Set<KType>,
         schemaObject: TopLevelObject?
     ): GraphQLSchema.Builder {
-        if (optInFederationV2) {
-            // preprocess any @LinkDirective applications to capture namespaces for all the imported specs
-            val appliedLinkDirectives = schemaObject?.kClass?.annotations?.filterIsInstance(LinkDirective::class.java)
-            appliedLinkDirectives?.forEach { appliedDirectiveAnnotation ->
-                val specUrl = Paths.get(appliedDirectiveAnnotation.url)
-                val spec = specUrl.parent.fileName.name
+        // preprocess any @LinkDirective applications to capture namespaces for all the imported specs
+        val appliedLinkDirectives = schemaObject?.kClass?.annotations?.filterIsInstance<LinkDirective>()
+        appliedLinkDirectives?.forEach { appliedDirectiveAnnotation ->
+            val specUrl = Paths.get(appliedDirectiveAnnotation.url)
+            val spec = specUrl.parent.fileName.name
 
-                if (linkSpecs.containsKey(spec)) {
-                    throw DuplicateSpecificationLinkImport(spec, appliedDirectiveAnnotation.url)
-                } else {
-                    val nameSpace: String = appliedDirectiveAnnotation.`as`.takeIf {
-                        it.isNotBlank()
-                    } ?: spec
-                    val imports: Map<String, String> = appliedDirectiveAnnotation.import.associate { import ->
-                        val importedName = import.`as`.takeIf { it.isNotBlank() } ?: import.name
-                        normalizeImportName(import.name) to normalizeImportName(importedName)
-                    }
-
-                    val linkSpec = LinkSpec(nameSpace, imports)
-                    linkSpecs[spec] = linkSpec
+            if (linkSpecs.containsKey(spec)) {
+                throw DuplicateSpecificationLinkImport(spec, appliedDirectiveAnnotation.url)
+            } else {
+                val nameSpace: String = appliedDirectiveAnnotation.`as`.takeIf {
+                    it.isNotBlank()
+                } ?: spec
+                val imports: Map<String, String> = appliedDirectiveAnnotation.import.associate { import ->
+                    val importedName = import.`as`.takeIf { it.isNotBlank() } ?: import.name
+                    normalizeImportName(import.name) to normalizeImportName(importedName)
                 }
-            }
 
-            // populate defaults
-            if (!linkSpecs.containsKey(FEDERATION_SPEC)) {
-                linkSpecs[FEDERATION_SPEC] = LinkSpec(
-                    FEDERATION_SPEC,
-                    listOf(
-                        COMPOSE_DIRECTIVE_NAME,
-                        EXTENDS_DIRECTIVE_NAME,
-                        EXTERNAL_DIRECTIVE_NAME,
-                        INACCESSIBLE_DIRECTIVE_NAME,
-                        INTERFACE_OBJECT_DIRECTIVE_NAME,
-                        KEY_DIRECTIVE_NAME,
-                        OVERRIDE_DIRECTIVE_NAME,
-                        PROVIDES_DIRECTIVE_NAME,
-                        REQUIRES_DIRECTIVE_NAME,
-                        SHAREABLE_DIRECTIVE_NAME,
-                        TAG_DIRECTIVE_NAME,
-                        FIELD_SET_SCALAR_NAME
-                    ).associateWith { it }
-                )
+                val linkSpec = LinkSpec(nameSpace, imports)
+                linkSpecs[spec] = linkSpec
             }
-            if (!linkSpecs.containsKey(LINK_SPEC)) {
-                linkSpecs[LINK_SPEC] = LinkSpec(LINK_SPEC, emptyMap())
-            }
+        }
+
+        // populate defaults
+        if (!linkSpecs.containsKey(FEDERATION_SPEC)) {
+            linkSpecs[FEDERATION_SPEC] = LinkSpec(
+                FEDERATION_SPEC,
+                listOf(
+                    COMPOSE_DIRECTIVE_NAME,
+                    EXTENDS_DIRECTIVE_NAME,
+                    EXTERNAL_DIRECTIVE_NAME,
+                    INACCESSIBLE_DIRECTIVE_NAME,
+                    INTERFACE_OBJECT_DIRECTIVE_NAME,
+                    KEY_DIRECTIVE_NAME,
+                    OVERRIDE_DIRECTIVE_NAME,
+                    PROVIDES_DIRECTIVE_NAME,
+                    REQUIRES_DIRECTIVE_NAME,
+                    SHAREABLE_DIRECTIVE_NAME,
+                    TAG_DIRECTIVE_NAME,
+                    FIELD_SET_SCALAR_NAME
+                ).associateWith { it }
+            )
+        }
+        if (!linkSpecs.containsKey(LINK_SPEC)) {
+            linkSpecs[LINK_SPEC] = LinkSpec(LINK_SPEC, emptyMap())
         }
 
         return super.willBuildSchema(queries, mutations, subscriptions, additionalTypes, additionalInputTypes, schemaObject)
@@ -249,31 +216,17 @@ open class FederatedSchemaGeneratorHooks(
     }
 
     override fun willGenerateDirective(directiveInfo: DirectiveMetaInformation): GraphQLDirective? =
-        if (optInFederationV2) {
-            willGenerateFederatedDirectiveV2(directiveInfo)
-        } else {
-            willGenerateFederatedDirective(directiveInfo)
+        when (directiveInfo.effectiveName) {
+            CONTACT_DIRECTIVE_NAME -> CONTACT_DIRECTIVE_TYPE
+            EXTERNAL_DIRECTIVE_NAME -> EXTERNAL_DIRECTIVE_TYPE
+            KEY_DIRECTIVE_NAME -> keyDirectiveDefinition(fieldSetScalar)
+            LINK_DIRECTIVE_NAME -> linkDirectiveDefinition(linkImportScalar)
+            POLICY_DIRECTIVE_NAME -> policyDirectiveDefinition(policiesScalar)
+            PROVIDES_DIRECTIVE_NAME -> providesDirectiveDefinition(fieldSetScalar)
+            REQUIRES_DIRECTIVE_NAME -> requiresDirectiveDefinition(fieldSetScalar)
+            REQUIRES_SCOPE_DIRECTIVE_NAME -> requiresScopesDirectiveType(scopesScalar)
+            else -> super.willGenerateDirective(directiveInfo)
         }
-
-    private fun willGenerateFederatedDirective(directiveInfo: DirectiveMetaInformation): GraphQLDirective? = when {
-        federationV2OnlyDirectiveNames.contains(directiveInfo.effectiveName) -> throw IncorrectFederatedDirectiveUsage(directiveInfo.effectiveName)
-        CONTACT_DIRECTIVE_NAME == directiveInfo.effectiveName -> CONTACT_DIRECTIVE_TYPE
-        EXTERNAL_DIRECTIVE_NAME == directiveInfo.effectiveName -> EXTERNAL_DIRECTIVE_TYPE
-        KEY_DIRECTIVE_NAME == directiveInfo.effectiveName -> KEY_DIRECTIVE_TYPE
-        else -> super.willGenerateDirective(directiveInfo)
-    }
-
-    private fun willGenerateFederatedDirectiveV2(directiveInfo: DirectiveMetaInformation): GraphQLDirective? = when (directiveInfo.effectiveName) {
-        CONTACT_DIRECTIVE_NAME -> CONTACT_DIRECTIVE_TYPE
-        EXTERNAL_DIRECTIVE_NAME -> EXTERNAL_DIRECTIVE_TYPE_V2
-        KEY_DIRECTIVE_NAME -> keyDirectiveDefinition(fieldSetScalar)
-        LINK_DIRECTIVE_NAME -> linkDirectiveDefinition(linkImportScalar)
-        POLICY_DIRECTIVE_NAME -> policyDirectiveDefinition(policiesScalar)
-        PROVIDES_DIRECTIVE_NAME -> providesDirectiveDefinition(fieldSetScalar)
-        REQUIRES_DIRECTIVE_NAME -> requiresDirectiveDefinition(fieldSetScalar)
-        REQUIRES_SCOPE_DIRECTIVE_NAME -> requiresScopesDirectiveType(scopesScalar)
-        else -> super.willGenerateDirective(directiveInfo)
-    }
 
     override fun willApplyDirective(directiveInfo: DirectiveMetaInformation, directive: GraphQLDirective): GraphQLAppliedDirective? {
         return when (directiveInfo.effectiveName) {
@@ -290,18 +243,16 @@ open class FederatedSchemaGeneratorHooks(
     }
 
     override fun didGenerateDirective(directiveInfo: DirectiveMetaInformation, directive: GraphQLDirective): GraphQLDirective {
-        if (optInFederationV2) {
-            // namespace generated directive if needed
-            val linkedSpec = directiveInfo.directive.annotationClass.annotations
-                .filterIsInstance(LinkedSpec::class.java)
-                .map { it.value }
-                .firstOrNull()
-            if (linkedSpec != null) {
-                val finalName = namespacedTypeName(linkedSpec, directive.name)
-                if (finalName != directive.name) {
-                    return directive.transform {
-                        it.name(finalName)
-                    }
+        // namespace generated directive if needed
+        val linkedSpec = directiveInfo.directive.annotationClass.annotations
+            .filterIsInstance<LinkedSpec>()
+            .map { it.value }
+            .firstOrNull()
+        if (linkedSpec != null) {
+            val finalName = namespacedTypeName(linkedSpec, directive.name)
+            if (finalName != directive.name) {
+                return directive.transform {
+                    it.name(finalName)
                 }
             }
         }
@@ -317,37 +268,32 @@ open class FederatedSchemaGeneratorHooks(
         val originalSchema = builder.build()
         val originalQuery = originalSchema.queryType
 
-        findMissingFederationDirectives(originalSchema.directives).forEach {
-            builder.additionalDirective(it)
+        // apply @link federation spec import only if it was not yet specified
+        val federationSpecImportExists = originalSchema.schemaAppliedDirectives.filter { it.name == "link" }.any {
+            it.getArgument("url")?.argumentValue?.value?.toString()?.startsWith(FEDERATION_SPEC_URL_PREFIX) == true
         }
-        if (optInFederationV2) {
-            // apply @link federation spec import only if it was not yet specified
-            val federationSpecImportExists = originalSchema.schemaAppliedDirectives.filter { it.name == "link" }.any {
-                it.getArgument("url")?.argumentValue?.value?.toString()?.startsWith(FEDERATION_SPEC_URL_PREFIX) == true
-            }
-            if (!federationSpecImportExists) {
-                val fed2Imports = linkSpecs[FEDERATION_SPEC]?.imports
-                    ?.keys
-                    ?.mapNotNull {
-                        val directive = originalSchema.getDirective(it)
-                        if (directive != null) {
-                            return@mapNotNull "@${directive.name}"
-                        }
-
-                        val scalar = originalSchema.getType(it) as? GraphQLNamedType
-                        if (scalar != null) {
-                            return@mapNotNull scalar.name
-                        }
-                        null
+        if (!federationSpecImportExists) {
+            val fed2Imports = linkSpecs[FEDERATION_SPEC]?.imports
+                ?.keys
+                ?.mapNotNull {
+                    val directive = originalSchema.getDirective(it)
+                    if (directive != null) {
+                        return@mapNotNull "@${directive.name}"
                     }
-                    ?: emptyList()
-                val linkDirective = linkDirectiveDefinition(linkImportScalar)
-                if (!originalSchema.directives.any { it.name == LINK_DIRECTIVE_NAME }) {
-                    // only add @link directive definition if it doesn't exist yet
-                    builder.additionalDirective(linkDirective)
+
+                    val scalar = originalSchema.getType(it) as? GraphQLNamedType
+                    if (scalar != null) {
+                        return@mapNotNull scalar.name
+                    }
+                    null
                 }
-                builder.withSchemaAppliedDirective(linkDirective.toAppliedLinkDirective(FEDERATION_SPEC_LATEST_URL, null, fed2Imports))
+                ?: emptyList()
+            val linkDirective = linkDirectiveDefinition(linkImportScalar)
+            if (!originalSchema.directives.any { it.name == LINK_DIRECTIVE_NAME }) {
+                // only add @link directive definition if it doesn't exist yet
+                builder.additionalDirective(linkDirective)
             }
+            builder.withSchemaAppliedDirective(linkDirective.toAppliedLinkDirective(FEDERATION_SPEC_LATEST_URL, null, fed2Imports))
         }
 
         val federatedCodeRegistry = GraphQLCodeRegistry.newCodeRegistry(originalSchema.codeRegistry)
@@ -367,28 +313,11 @@ open class FederatedSchemaGeneratorHooks(
                 .additionalType(ANY_SCALAR_TYPE)
         }
 
-        val federatedBuilder = if (optInFederationV2) {
-            builder
-        } else {
-            // transform schema to rename FieldSet to _FieldSet
-            GraphQLSchema.newSchema(SchemaTransformer.transformSchema(builder.build(), FieldSetTransformer()))
-        }
-
         // Register the data fetcher for the _service query
-        val sdl = getFederatedServiceSdl(federatedBuilder.build())
+        val sdl = getFederatedServiceSdl(builder.build())
         federatedCodeRegistry.dataFetcher(FieldCoordinates.coordinates(originalQuery.name, SERVICE_FIELD_DEFINITION.name), DataFetcher { _Service(sdl) })
 
-        return federatedBuilder.codeRegistry(federatedCodeRegistry.build())
-    }
-
-    private fun findMissingFederationDirectives(existingDirectives: List<GraphQLDirective>): List<GraphQLDirective> = if (optInFederationV2) {
-        emptyList()
-    } else {
-        // we auto-add directive definitions only for fed v1 schemas
-        val existingDirectiveNames = existingDirectives.map { it.name }
-        federatedDirectiveV1List.filter {
-            !existingDirectiveNames.contains(it.name)
-        }
+        return builder.codeRegistry(federatedCodeRegistry.build())
     }
 
     /**
@@ -399,11 +328,6 @@ open class FederatedSchemaGeneratorHooks(
      */
     override fun didGenerateQueryObject(type: GraphQLObjectType): GraphQLObjectType = GraphQLObjectType.newObject(type)
         .field(SERVICE_FIELD_DEFINITION)
-        .also {
-            if (!optInFederationV2) {
-                it.withAppliedDirective(EXTENDS_DIRECTIVE_TYPE.toAppliedDirective())
-            }
-        }
         .build()
 
     /**
@@ -412,13 +336,7 @@ open class FederatedSchemaGeneratorHooks(
      * See the federation spec for more details:
      * https://www.apollographql.com/docs/federation/subgraph-spec/#enhanced-introspection-with-query_service
      */
-    private fun getFederatedServiceSdl(schema: GraphQLSchema): String {
-        return if (optInFederationV2) {
-            generateServiceSDLV2(schema)
-        } else {
-            generateServiceSDL(schema, false)
-        }
-    }
+    private fun getFederatedServiceSdl(schema: GraphQLSchema): String = generateServiceSDLV2(schema)
 
     /**
      * Get all the federation entities in the _Entity union, aka all the types with the @key directive.
@@ -427,11 +345,8 @@ open class FederatedSchemaGeneratorHooks(
      * https://www.apollographql.com/docs/apollo-server/federation/federation-spec/#union-_entity
      */
     private fun getFederatedEntities(originalSchema: GraphQLSchema): Set<String> {
-        val keyDirectiveName = if (optInFederationV2) {
-            namespacedTypeName(FEDERATION_SPEC, KEY_DIRECTIVE_NAME)
-        } else {
-            KEY_DIRECTIVE_NAME
-        }
+        val keyDirectiveName = namespacedTypeName(FEDERATION_SPEC, KEY_DIRECTIVE_NAME)
+
         val entities = originalSchema.allTypesAsList
             .filterIsInstance<GraphQLDirectiveContainer>()
             .filter { type -> type.hasAppliedDirective(keyDirectiveName) }
