+++ generator/graphql-kotlin-federation/src/test/kotlin/com/expediagroup/graphql/generator/federation/data/TestResolvers.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Expedia, Inc
+ * Copyright 2024 Expedia, Inc
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,9 +16,9 @@
 
 package com.expediagroup.graphql.generator.federation.data
 
-import com.expediagroup.graphql.generator.federation.data.queries.federated.v1.Author
-import com.expediagroup.graphql.generator.federation.data.queries.federated.v1.Book
-import com.expediagroup.graphql.generator.federation.data.queries.federated.v1.User
+import com.expediagroup.graphql.generator.federation.data.queries.federated.Author
+import com.expediagroup.graphql.generator.federation.data.queries.federated.Book
+import com.expediagroup.graphql.generator.federation.data.queries.federated.User
 import com.expediagroup.graphql.generator.federation.execution.FederatedTypePromiseResolver
 import com.expediagroup.graphql.generator.federation.execution.FederatedTypeSuspendResolver
 import graphql.schema.DataFetchingEnvironment
