+++ generator/graphql-kotlin-federation/src/test/kotlin/com/expediagroup/graphql/generator/federation/execution/EntitiesDataFetcherTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Expedia, Inc
+ * Copyright 2024 Expedia, Inc
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,9 +19,9 @@ package com.expediagroup.graphql.generator.federation.execution
 import com.expediagroup.graphql.generator.federation.data.AuthorResolver
 import com.expediagroup.graphql.generator.federation.data.BookResolver
 import com.expediagroup.graphql.generator.federation.data.UserResolver
-import com.expediagroup.graphql.generator.federation.data.queries.federated.v1.Author
-import com.expediagroup.graphql.generator.federation.data.queries.federated.v1.Book
-import com.expediagroup.graphql.generator.federation.data.queries.federated.v1.User
+import com.expediagroup.graphql.generator.federation.data.queries.federated.Author
+import com.expediagroup.graphql.generator.federation.data.queries.federated.Book
+import com.expediagroup.graphql.generator.federation.data.queries.federated.User
 import graphql.GraphQLContext
 import graphql.GraphQLError
 import graphql.schema.DataFetchingEnvironment
