+++ generator/graphql-kotlin-federation/src/test/kotlin/com/expediagroup/graphql/generator/federation/execution/ServiceQueryResolverTest.kt
@@ -19,7 +19,7 @@ package com.expediagroup.graphql.generator.federation.execution
 import com.expediagroup.graphql.generator.TopLevelObject
 import com.expediagroup.graphql.generator.federation.FederatedSchemaGeneratorConfig
 import com.expediagroup.graphql.generator.federation.FederatedSchemaGeneratorHooks
-import com.expediagroup.graphql.generator.federation.data.queries.federated.v1.CustomScalar
+import com.expediagroup.graphql.generator.federation.data.queries.federated.CustomScalar
 import com.expediagroup.graphql.generator.federation.data.queries.simple.NestedQuery
 import com.expediagroup.graphql.generator.federation.data.queries.simple.SimpleQuery
 import com.expediagroup.graphql.generator.federation.toFederatedSchema
@@ -39,54 +39,6 @@ import kotlin.reflect.KClass
 import kotlin.reflect.KType
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
-
-// SDL is returned without _entity and _service queries
-const val FEDERATED_SERVICE_SDL =
-"""
-schema {
-  query: Query
-}
-
-interface Product @extends @key(fields : "id") @key(fields : "upc") {
-  id: String! @external
-  reviews: [Review!]!
-  upc: String! @external
-}
-
-type Author @extends @key(fields : "authorId") {
-  authorId: Int! @external
-  name: String! @external
-}
-
-type Book implements Product @extends @key(fields : "id") @key(fields : "upc") {
-  author: User! @provides(fields : "name")
-  id: String! @external
-  reviews: [Review!]!
-  shippingCost: String! @requires(fields : "weight")
-  upc: String! @external
-  weight: Float! @external
-}
-
-type Query @extends
-
-type Review {
-  body: String!
-  content: String
-  customScalar: CustomScalar!
-  id: String!
-}
-
-type User @extends @key(fields : "userId") {
-  name: String! @external
-  userId: Int! @external
-}
-
-""${'"'}
-This is a multi-line comment on a custom scalar.
-This should still work multiline and double quotes (") in the description.
-Line 3.
-""${'"'}
-scalar CustomScalar"""
 
 const val BASE_SERVICE_SDL =
 """
@@ -198,7 +150,7 @@ scalar link__Import
 
 class ServiceQueryResolverTest {
 
-    class CustomScalarFederatedHooks : FederatedSchemaGeneratorHooks(emptyList(), optInFederationV2 = false) {
+    class CustomScalarFederatedHooks : FederatedSchemaGeneratorHooks(emptyList()) {
         override fun willGenerateGraphQLType(type: KType): GraphQLType? = when (type.classifier as? KClass<*>) {
             CustomScalar::class -> graphqlCustomScalar
             else -> super.willGenerateGraphQLType(type)
@@ -226,36 +178,6 @@ class ServiceQueryResolverTest {
             override fun parseLiteral(input: Value<*>, variables: CoercedVariables, graphQLContext: GraphQLContext, locale: Locale): CustomScalar {
                 val customValue = (input as? StringValue)?.value ?: throw CoercingParseValueException("Cannot parse $input to CustomScalar")
                 return CustomScalar(customValue)
-            }
-        }
-    }
-
-    @Test
-    fun `verify can retrieve SDL using _service query`() {
-        val config = FederatedSchemaGeneratorConfig(
-            supportedPackages = listOf("com.expediagroup.graphql.generator.federation.data.queries.federated.v1"),
-            hooks = CustomScalarFederatedHooks()
-        )
-
-        val schema = toFederatedSchema(config = config)
-        val query =
-            """
-                query sdlQuery {
-                  _service {
-                    sdl
-                  }
-                }
-            """.trimIndent()
-        val executionInput = ExecutionInput.newExecutionInput()
-            .query(query)
-            .build()
-        val graphQL = GraphQL.newGraphQL(schema).build()
-        val result = graphQL.executeAsync(executionInput).get().toSpecification()
-
-        assertNotNull(result["data"] as? Map<*, *>) { data ->
-            assertNotNull(data["_service"] as? Map<*, *>) { queryResult ->
-                val sdl = queryResult["sdl"] as? String
-                assertEquals(FEDERATED_SERVICE_SDL.trim(), sdl)
             }
         }
     }
@@ -293,7 +215,7 @@ class ServiceQueryResolverTest {
     @Test
     fun `verify can retrieve Federation v2 SDL using _service query`() {
         val config = FederatedSchemaGeneratorConfig(
-            supportedPackages = listOf("com.expediagroup.graphql.generator.federation.data.queries.federated.v2"),
+            supportedPackages = listOf("com.expediagroup.graphql.generator.federation.data.queries.federated"),
             hooks = FederatedSchemaGeneratorHooks(emptyList())
         )
