+++ servers/graphql-kotlin-spring-server/src/main/kotlin/com/expediagroup/graphql/server/spring/FederatedSchemaAutoConfiguration.kt
@@ -61,7 +61,6 @@ class FederatedSchemaAutoConfiguration(
         resolvers: Optional<List<FederatedTypeResolver>>
     ): FederatedSchemaGeneratorHooks = FederatedSchemaGeneratorHooks(
         resolvers.orElse(emptyList()),
-        config.federation.optInV2
     )
 
     @Bean
