+++ website/docs/schema-generator/federation/apollo-federation.mdx
@@ -19,34 +19,6 @@ underlying graph and convey the relationships between different schema types. Ea
 valid GraphQL schema and can be run independently. This is in contrast with a traditional schema stitching approach where
 relationships between individual services, i.e. linking configuration, is configured at the GraphQL gateway level.
 
-## Federation v1 vs Federation v2
-
-Federation v2 is an evolution of the Federation spec to make it more powerful, flexible and easier to adapt. While v1 and
-v2 schemas are similar in many ways, Federation v2 relaxes some of the constraints and adds additional capabilities. See
-[Apollo documentation](https://www.apollographql.com/docs/federation/federation-2/new-in-federation-2/) for details.
-
-By default, `graphql-kotlin-federation` library will generate Federation v2 compatible schema. In order to generate v1
-compatible schema you have to explicitly opt-out by specifying `optInFederationV2 = false` on your instance of
-`FederatedSchemaGeneratorHooks`.
-
-```kotlin
-val myHooks = FederatedSchemaGeneratorHooks(resolvers = myFederatedResolvers)
-val myConfig = FederatedSchemaGeneratorConfig(
-  supportedPackages = "com.example",
-  hooks = myHooks
-)
-
-toFederatedSchema(
-  config = myConfig,
-  queries = listOf(TopLevelObject(MyQuery()))
-)
-```
-
-:::note
-When generating federated schemas, `graphql-kotlin-spring-server` defaults to Federation v2. If you want to generate Federation
-v1 schema, you have to explicitly opt-out by configuring `graphql.federation.optInV2 = false` property.
-:::
-
 ## Install
 
 Using a JVM dependency manager, link `graphql-kotlin-federation` to your project.
@@ -124,7 +96,7 @@ schema @link(import : ["@key", "FieldSet"], url : "https://specs.apollo.dev/fede
 }
 
 directive @key(fields: FieldSet!, resolvable: Boolean = true) repeatable on OBJECT | INTERFACE
-directive @link(import: [String], url: String!) repeatable on SCHEMA
+directive @link(as: String, import: [link__Import], url: String!) repeatable on SCHEMA
 
 type Query {
    getUsers: [User!]!
@@ -146,4 +118,5 @@ type _Service {
 
 scalar FieldSet
 scalar _Any
+scalar link__Import
 ```
