+++ website/docs/schema-generator/federation/federated-schemas.md
@@ -15,20 +15,22 @@ a way to instantiate the underlying federated objects by implementing correspond
 is then generated by invoking the `toFederatedSchema` function
 ([link](https://github.com/ExpediaGroup/graphql-kotlin/blob/master/generator/graphql-kotlin-federation/src/main/kotlin/com/expediagroup/graphql/generator/federation/toFederatedSchema.kt#L34)).
 
-**In order to generate valid federated schemas, you will need to annotate both your base schema and the one extending
-it**. Federated Gateway (e.g. Apollo) will then combine the individual graphs to form single federated graph.
+**In order to generate valid federated schemas, you will need to annotate your entities in all your subgraphs**.
+Federated Gateway (e.g. Apollo) will then combine the individual graphs to form single federated graph.
 
 :::caution
 If you are using custom `Query` type then all of you federated GraphQL services have to use the same type. It is
 not possible for federated services to have different definitions of `Query` type.
 :::
 
-### Base Subgraph
+### Subgraph A
 
-Base schema defines GraphQL types that will be extended by schemas exposed by other GraphQL services. In the example
-below, we define base `Product` type with `id` and `description` fields. `id` is the primary key that uniquely
-identifies the `Product` type object and is specified in `@key` directive. Since it is a base schema that doesn't expose
-any extended functionality our FederatedTypeRegistry does not include any federated resolvers.
+Federation v2 relaxed entity ownership and now every subgraph that defines given entity is its owner. In the example
+below, we define `Product` type with `id` and `description` fields. `id` is the primary key that uniquely
+identifies the `Product` type object and is specified in `@key` directive. Since it might be possible to resolve
+`Product` entity from other subgraphs, we also should specify an "entry point" for the federated type - we need to
+create a `FederatedTypeResolver` that will be used to instantiate the federated `Product` type when processing federated
+queries.
 
 ```kotlin
 @KeyDirective(fields = FieldSet("id"))
@@ -40,8 +42,20 @@ class ProductQuery {
   }
 }
 
+// Resolve a "Product" type from the _entities query
+class ProductResolver : FederatedTypeSuspendResolver<Product> {
+    override val typeName = "Product"
+
+    override suspend fun resolve(
+        environment: DataFetchingEnvironment,
+        representation: Map<String, Any>
+    ): Product? =
+        representation["id"]?.toString()?.toIntOrNull()?.let { id -> Product(id) }
+}
+
 // Generate the schema
-val hooks = FederatedSchemaGeneratorHooks(emptyList())
+val resolvers = listOf(ProductResolver())
+val hooks = FederatedSchemaGeneratorHooks(resolvers)
 val config = FederatedSchemaGeneratorConfig(supportedPackages = listOf("org.example"), hooks = hooks)
 val queries = listOf(TopLevelObject(ProductQuery()))
 
@@ -73,20 +87,16 @@ type _Service {
 }
 ```
 
-### Extended Subgraph
+### Subgraph B
 
-Extended federated GraphQL schemas provide additional functionality to the types already exposed by other GraphQL
-services. In the example below, `Product` type is extended to add new `reviews` field to it. Primary key needed to
-instantiate the `Product` type (i.e. `id`) has to match the `@key` definition on the base type. Since primary keys are
-defined on the base type and are only referenced from the extended type, all of the fields that are part of the field
-set specified in `@key` directive have to be marked as `@external`. Finally, we also need to specify an "entry point"
-for the federated type - we need to create a FederatedTypeResolver that will be used to instantiate the federated
-`Product` type when processing federated queries.
+Each subgraph can extend and provide new functionality to entities defined in other subgraphs. In the example below,
+`Product` type is extended to add new `reviews` field to it. Primary key needed to instantiate the `Product` type (i.e. `id`)
+has to match one of the entity `@key` definitions defined in other subgraphs. Finally, we also need to specify an "entry point"
+for the federated type - we need to create a FederatedTypeResolver.
 
 ```kotlin
 @KeyDirective(fields = FieldSet("id"))
-@ExtendsDirective
-data class Product(@ExternalDirective val id: Int) {
+data class Product(val id: Int) {
     // Add the "reviews" field to the type
     suspend fun reviews(): List<Review> = getReviewByProductId(id)
 }
@@ -143,7 +153,7 @@ type _Service {
 
 ### Federated Supergraph
 
-Once we have both base and extended GraphQL services up and running, we will also need to configure Federated Gateway
+Once we have both GraphQL subgraphs up and running, we will also need to configure Federated Gateway
 to combine them into a single supergraph schema. Using the examples above, our final federated schema will be generated as:
 
 ```graphql
