+++ website/docs/server/spring-server/spring-properties.md
@@ -16,7 +16,6 @@ details on the supported configuration properties.
 | graphql.packages                            | List of supported packages that can contain GraphQL schema type definitions                                                      |                               |
 | graphql.printSchema                         | Boolean flag indicating whether to print the schema after generator creates it                                                   | false                         |
 | graphql.federation.enabled                  | Boolean flag indicating whether to generate federated GraphQL model                                                              | false                         |
-| graphql.federation.optInV2                  | Boolean flag indicating whether to generate Federation v2 GraphQL model                                                          | false                         |
 | graphql.federation.tracing.enabled          | Boolean flag indicating whether add federated tracing data to the extensions                                                     | true (if federation enabled)  |
 | graphql.federation.tracing.debug            | Boolean flag to log debug info in the federated tracing                                                                          | false (if federation enabled) |
 | graphql.introspection.enabled               | Boolean flag indicating whether introspection queries are enabled                                                                | true                          |
